refactor(agents): share model-provider label helpers; trim duplicate comment
Self-review cleanup, no behavior change:
- Extract MODEL_PROVIDER_LABELS + modelProviderKey/modelOptionLabel into
  the shared schemaProperties module; desktop NewSessionView and mobile
  SchemaArgsControls now import them instead of each holding a byte-
  identical copy. Finishes the "extract rather than duplicate" approach
  the PR already applies to the classification helpers and removes the
  provider-label drift risk across the RN/DOM boundary.
- Trim the redundant "RN FormData lacks set" rationale in sendMessage so
  it lives once, at the addField branch where the append-vs-set choice is.
